requirements.txt
absl-py==2.4.0
altair==6.0.0
astunparse==1.6.3
attrs==25.4.0
blinker==1.9.0
cachetools==6.2.6
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
flatbuffers==25.12.19
fonttools==4.61.1
gast==0.7.0
gitdb==4.0.12
GitPython==3.1.46
google-pasta==0.2.0
grpcio==1.78.0
h5py==3.15.1
idna==3.11
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
keras==3.12.1
kiwisolver==1.4.9
libclang==18.1.1
Markdown==3.10.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
ml_dtypes==0.5.4
namex==0.1.0
narwhals==2.16.0
numpy==2.2.6
opencv-python==4.13.0.92
opt_einsum==3.4.0
optree==0.18.0
packaging==26.0
pandas==2.3.3
pillow==12.1.0
protobuf==6.33.5
pyarrow==23.0.0
pydeck==0.9.1
Pygments==2.19.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
pytz==2025.2
referencing==0.37.0
requests==2.32.5
rich==14.3.2
rpds-py==0.30.0
scikit-learn==1.7.2
scipy==1.15.3
six==1.17.0
smmap==5.0.2
streamlit==1.54.0
tenacity==9.1.4
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.3.0
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.5.4
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
watchdog==6.0.0
Werkzeug==3.1.5
wrapt==2.1.1
